// src/types/Errors.ts
export enum Errors {
  NONE = '',
  LOAD = 'Unable to load a user',
  REGISTRATION = 'Unable to registrate a user',
  ADD = 'Unable to add a todo',
  DELETE = 'Unable to delete a todo',
  UPDATE = 'Unable to update a todo',
  TITLE = "Title can't be empty",
  UNEXPECTED = 'Something went wrong. Unexpected error',
}